Часть правок
