+++ js/util.js
@@ -1,5 +1,3 @@
-const ALERT_SHOW_TIME = 5000;
-
 function getRandomNumber(min, max) {
   const minNumber = Math.ceil(min);
   const maxNumber = Math.floor(max);
@@ -40,71 +38,5 @@ function getRandomArray(list) {
 function getRandomElement(list) {
   return list[getRandomNumber(0, list.length - 1)];
 }
-const textAlert = 'Не удалось получить объявления. Попробуйте ещё раз';
 
-const showAlert = (message) => {
-  const alertContainer = document.createElement('div');
-  alertContainer.style.zIndex = 100;
-  alertContainer.style.position = 'absolute';
-  alertContainer.style.left = 0;
-  alertContainer.style.top = 0;
-  alertContainer.style.right = 0;
-  alertContainer.style.padding = '10px 3px';
-  alertContainer.style.fontSize = '30px';
-  alertContainer.style.textAlign = 'center';
-  alertContainer.style.backgroundColor = 'red';
-  alertContainer.textContent = message;
-  document.body.append(alertContainer);
-
-  setTimeout(() => {
-    alertContainer.remove();
-  }, ALERT_SHOW_TIME);
-};
-
-const showSuccess = () => {
-  const buttonEsc = 'Escape';
-  const buttonEsc2 = 27;
-  const successTemplate = document.querySelector('#success').content.querySelector('.success');
-  const successContainer = successTemplate.cloneNode(true);
-  document.body.append(successContainer);
-
-  document.body.addEventListener('keydown', (evt) => {
-    if (evt.key === buttonEsc || evt.key === buttonEsc2) {
-      successContainer.remove();
-      document.body.removeEventListener('keydown',successContainer);
-    }
-  });
-
-  document.body.addEventListener('click', () => {
-    successContainer.remove();
-    document.body.removeEventListener('click',successContainer);
-  });
-};
-
-const showError = () => {
-  const buttonEsc = 'Escape';
-  const buttonEsc2 = 27;
-  const errorTemplate = document.querySelector('#error').content.querySelector('.error');
-  const errorContainer = errorTemplate.cloneNode(true);
-  const errorButton = errorTemplate.querySelector('.error__button');
-  document.body.append(errorContainer);
-
-  document.body.addEventListener('keydown', (evt) => {
-    if (evt.key === buttonEsc || evt.key === buttonEsc2) {
-      errorContainer.remove();
-      document.body.removeEventListener('keydown',errorContainer);
-    }
-  });
-
-  document.body.addEventListener('click', () => {
-    errorContainer.remove();
-    document.body.removeEventListener('click',errorContainer);
-  });
-
-  errorButton.addEventListener('click', () => {
-    errorContainer.remove();
-    errorButton.removeEventListener('click',errorContainer);
-  });
-};
-
-export {getRandomNumber, getRandomFloat,getRandomArray, getRandomElement, showAlert,showSuccess, showError,textAlert};
+export {getRandomNumber, getRandomFloat,getRandomArray, getRandomElement};
